Allow to initialize environment
This just adds a way to define a function in the configuration which will receive the Environment. This makes it possible, for example, to define custom helpers, or to implement the default asset path helper of the Mincer context. Without this, the preprocessor cannot be used if your assets use asset path or other helpers.